moved npp and latencyhiding to benchmark folder
