+++ include/CMakeLists.txt
@@ -12,14 +12,6 @@ add_library(headers INTERFACE ${CUDA_SOURCES} )
 source_group(TREE ${CMAKE_CURRENT_SOURCE_DIR} FILES ${CUDA_SOURCES})
 #target_include_directories(include INTERFACE .)
 set_target_properties(${cuda_target} PROPERTIES CXX_STANDARD 17 CXX_STANDARD_REQUIRED YES CXX_EXTENSIONS NO)
-
  
-        #intellisense hack for vs2022
-        set (LAUNCH_SOURCES ${CMAKE_SOURCE_DIR}/include/intellisense/main.cpp ${CMAKE_SOURCE_DIR}/include/intellisense/main.h)
- 
-
-if (${BUILD_UTEST})       
-    discover_tests("${CMAKE_CURRENT_SOURCE_DIR}/tests")
-endif()
