+++ include/testing/fkbenchmark.h
@@ -1,34 +1,3 @@
-/* Copyright 2024 Oscar Amoros Huguet
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License. */
-
-#include <sstream>
-#include <fstream>
-#include <iostream>
-#include <unordered_map>
-#include <array>
-
-#include <fused_kernel/core/utils/vlimits.h>
-
-template <size_t START_VALUE, size_t INCREMENT, std::size_t... Is>
-constexpr std::array<size_t, sizeof...(Is)> generate_sequence(std::index_sequence<Is...>) {
-    return std::array<size_t, sizeof...(Is)>{(START_VALUE + (INCREMENT * Is))...};
-}
-
-template <size_t START_VALUE, size_t INCREMENT, size_t NUM_ELEMS>
-constexpr std::array<size_t, NUM_ELEMS> arrayIndexSecuence = generate_sequence<START_VALUE, INCREMENT>(std::make_index_sequence<NUM_ELEMS>{});
-
-#ifdef ENABLE_BENCHMARK
 std::unordered_map<std::string, std::stringstream> benchmarkResultsText;
 std::unordered_map<std::string, std::ofstream> currentFile;
 // Select the path where to write the benchmark files
@@ -84,10 +53,8 @@ inline void processExecution(const BenchmarkResultsNumbers& resF,
         currentFile[fileName] << std::endl;
     }
 }
-#endif
-
-#ifdef ENABLE_BENCHMARK
-#define START_CVGS_BENCHMARK \
+  
+#define START_FK_BENCHMARK \
 std::cout << "Executing " << __func__ << " fusing " << VARIABLE_DIMENSION << " operations. " << std::endl; \
 cudaEvent_t start, stop; \
 BenchmarkResultsNumbers resF; \
@@ -99,12 +66,8 @@ gpuErrchk(cudaEventCreate(&stop)); \
 std::array<float, ITERS> cvGSelapsedTime; \
 for (int i = 0; i < ITERS; i++) { \
 gpuErrchk(cudaEventRecord(start, stream));
-#else
-#define START_CVGS_BENCHMARK
-#endif
-
-#ifdef ENABLE_BENCHMARK
-#define STOP_CVGS_BENCHMARK \
+ 
+#define STOP_FK_BENCHMARK \
 gpuErrchk(cudaEventRecord(stop, stream)); \
 gpuErrchk(cudaEventSynchronize(stop)); \
 gpuErrchk(cudaEventElapsedTime(&cvGSelapsedTime[i], start, stop)); \
@@ -113,15 +76,10 @@ resF.cvGSelapsedTimeMin = resF.cvGSelapsedTimeMin > cvGSelapsedTime[i] ? cvGSela
 resF.cvGSelapsedTimeAcum += cvGSelapsedTime[i]; \
 } \
 processExecution<VARIABLE_DIMENSION, ITERS, variableDimanesionValues.size(), variableDimanesionValues>(resF, __func__, cvGSelapsedTime, VARIABLE_DIMENSION_NAME);
-#else
-#define STOP_CVGS_BENCHMARK
-#endif
-
-#ifdef ENABLE_BENCHMARK
+ 
 #define CLOSE_BENCHMARK \
 for (auto&& [_, file] : currentFile) { \
     file.close(); \
 }
-#else
-#define CLOSE_BENCHMARK
-#endif
+ 
+
