+++ include/testing/nppbenchmark.h
@@ -1,34 +1,3 @@
-/* Copyright 2024 Albert Andaluz Gonzalez
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License. */
-
- 
-#include <fused_kernel/core/utils/vlimits.h>
-
-#include <array>
-#include <fstream>
-#include <sstream>
-#include <unordered_map>
-
-template <size_t START_VALUE, size_t INCREMENT, std::size_t... Is>
-constexpr std::array<size_t, sizeof...(Is)> generate_sequence(std::index_sequence<Is...>) {
-  return std::array<size_t, sizeof...(Is)>{(START_VALUE + (INCREMENT * Is))...};
-}
-
-template <size_t START_VALUE, size_t INCREMENT, size_t NUM_ELEMS>
-constexpr std::array<size_t, NUM_ELEMS> arrayIndexSecuence =
-    generate_sequence<START_VALUE, INCREMENT>(std::make_index_sequence<NUM_ELEMS>{});
- 
 #ifdef ENABLE_BENCHMARK
 std::unordered_map<std::string, std::stringstream> benchmarkResultsText;
 std::unordered_map<std::string, std::ofstream> currentFile;
