+++ tests/algorithm/test_crop.cu
@@ -12,7 +12,7 @@
    See the License for the specific language governing permissions and
    limitations under the License. */
 
-#include "tests/main.h"
+#include <testing/main.h>
 
 #include <fused_kernel/algorithms/image_processing/crop.cuh>
 #include <fused_kernel/core/data/ptr_nd.cuh>
