+++ tests/cudabug/test_nvcc14_15_BugReproducer.cu
@@ -13,7 +13,7 @@
    limitations under the License. */
 
 
-#include "tests/main.h"
+#include <testing/main.h>
 
 #include <type_traits>
