+++ tests/data/test_ptr_nd.cu
@@ -14,7 +14,7 @@
 
 
 #ifdef WIN32
-#include "tests/main.h"
+#include <testing/main.h>
 #endif
 
 #include <fused_kernel/core/data/ptr_nd.cuh>
