+++ tests/data/test_rect.cu
@@ -14,7 +14,7 @@
 
 
 #ifdef WIN32
-#include "tests/main.h"
+#include <testing/main.h>
 #endif
 
 #include <fused_kernel/core/data/rect.h>
