+++ tests/device/test_device_functions.cu
@@ -13,7 +13,7 @@
    limitations under the License. */
 
 
-#include "tests/main.h"
+#include <testing/main.h>
 
 #include <fused_kernel/algorithms/basic_ops/arithmetic.cuh>
 #include <fused_kernel/algorithms/basic_ops/cast.cuh>
