+++ tests/operation/test_fused_operation.cu
@@ -13,7 +13,7 @@
    limitations under the License. */
 
 
-#include "tests/main.h"
+#include <testing/main.h>
 
 #include <fused_kernel/algorithms/basic_ops/arithmetic.cuh>
 #include <fused_kernel/algorithms/basic_ops/cast.cuh>
