+++ tests/operation/test_operation_tuple.cu
@@ -13,7 +13,7 @@
    limitations under the License. */
 
 
-#include "tests/main.h"
+#include <testing/main.h>
 
 #include <fused_kernel/core/data/ptr_nd.cuh>
 #include <fused_kernel/core/execution_model/operation_tuple.cuh>
