+++ tests/operation/test_operation_types.cu
@@ -12,7 +12,7 @@
    See the License for the specific language governing permissions and
    limitations under the License. */
 
-#include "tests/main.h"
+#include <testing/main.h>
  
 
 #include <fused_kernel/algorithms/basic_ops/arithmetic.cuh>
